// src/types.ts
export enum HookType {
    CUSTOM_CONTRACT = 'CUSTOM_CONTRACT',
    PASSWORD = 'PASSWORD',
    PROMOCODE = 'PROMOCODE',
    CAPTCHA = 'CAPTCHA',
    GUILD = 'GUILD',
}

export const HooksName = [
    'onKeyPurchaseHook',
    'onKeyCancelHook',
    'onValidKeyHook',
    'onTokenURIHook',
    'onKeyTransferHook',
    'onKeyExtendHook',
    'onKeyGrantHook',
] as const

export type HookName = (typeof HooksName)[number]

export interface Hook {
    id: HookType
    name: string
    address: string
    description?: string
}